Adding DAO interfaces layer.

+++ src/main/java/com/company/stockservice/model/MutableProductCatalogItem.java
@@ -1,8 +1,14 @@
 package com.company.stockservice.model;
 
+/**
+ * This interface ensures modification on ProductCatalogItem object
+ * 
+ * @author marcin
+ *
+ */
 public interface MutableProductCatalogItem extends ProductCatalogItem {
 
 	public void setProduct(Product product);
 
-	public void setAmount(Integer amount);
+	public void setAmount(Double amount);
 }
