+++ src/main/java/com/company/stockservice/model/ProductCatalogItem.java
@@ -1,8 +1,14 @@
 package com.company.stockservice.model;
 
+/**
+ * Model class which defines ProductCatalogItem entity
+ * 
+ * @author Malcinn
+ *
+ */
 public interface ProductCatalogItem {
 
 	public Product getProduct();
 
-	public Integer getAmount();
+	public Double getAmount();
 }
